refactor: add correct folder names to namespaces
